MBVM-52: Load search indexes from FTP
Reindexing the whole database from scratch can take a very long time.

This patch helps with loading search indexes from FTP instead.

Loading search indexes takes two steps and requires under 80GB:

1. Fetch dump files from FTP, it requires under 30GB.
   Dump files are saved in the volume `searchdump`.

2. Extract dump files in place of Solr data, it requires under 50GB.
   The Docker image tag `metabrainz/mb-solr:3.1.3` comes with `zstd`.

+++ build/musicbrainz/scripts/fetch-dump.sh
@@ -3,6 +3,7 @@
 set -e -o pipefail -u
 
 DB_DUMP_DIR=/media/dbdump
+SEARCH_DUMP_DIR=/media/searchdump
 BASE_FTP_URL='ftp://ftp.eu.metabrainz.org/pub/musicbrainz'
 TARGET=''
 WGET_CMD=(wget)
@@ -11,11 +12,13 @@ SCRIPT_NAME=$(basename "$0")
 HELP=$(cat <<EOH
 Usage: $SCRIPT_NAME [<options>] <target>
 
-Fetch dump files of the MusicBrainz database.
+Fetch dump files of the MusicBrainz database and/or search indexes.
 
 Targets:
+  both          Fetch latest search dump with replica dump of the same day.
   replica       Fetch latest database's replicated tables only.
   sample        Fetch latest database's sample only.
+  search        Fetch latest search indexes only.
 
 Options:
   --base-ftp-url <url>          Specify URL to MetaBrainz/MusicBrainz FTP directory.
@@ -34,7 +37,7 @@ EOH
 while [[ $# -gt 0 ]]
 do
 	case "$1" in
-		replica | sample )
+		both | replica | sample | search )
 			if [[ -n $TARGET ]]
 			then
 				echo >&2 "$SCRIPT_NAME: only one target argument can be given"
@@ -78,14 +81,36 @@ then
 	exit 64 # EX_USAGE
 fi
 
+# Fetch latest search indexes
+
+if [[ $TARGET =~ ^(both|search)$ ]]
+then
+	echo "$(date): Fetching search indexes dump..."
+	cd "$SEARCH_DUMP_DIR" && find . -delete && cd -
+	"${WGET_CMD[@]}" -nd -nH -P "$SEARCH_DUMP_DIR" \
+		"$BASE_FTP_URL/data/search-indexes/LATEST"
+	DUMP_TIMESTAMP=$(cat /media/searchdump/LATEST)
+	"${WGET_CMD[@]}" -nd -nH -r -P "$SEARCH_DUMP_DIR" \
+		"$BASE_FTP_URL/data/search-indexes/$DUMP_TIMESTAMP/"
+	cd "$SEARCH_DUMP_DIR" && md5sum -c MD5SUMS && cd -
+	if [[ $TARGET == search ]]
+	then
+		echo 'Done fetching search indexes dump'
+		exit 0 # EX_OK
+	fi
+fi
+
 # Prepare to fetch database dump
 
-echo "$(date): Fetching database dump..."
+if [[ $TARGET != search ]]
+then
+	echo "$(date): Fetching database dump..."
 
-rm -rf "${DB_DUMP_DIR:?}"/*
+	rm -rf "${DB_DUMP_DIR:?}"/*
+fi
 
 case "$TARGET" in
-	replica )
+	both | replica )
 		DB_DUMP_REMOTE_DIR=data/fullexport
 		DB_DUMP_FILES=(
 			mbdump.tar.bz2
@@ -104,13 +129,31 @@ case "$TARGET" in
 		;;
 esac
 
-"${WGET_CMD[@]}" -nd -nH -P "$DB_DUMP_DIR" \
-	"$BASE_FTP_URL/$DB_DUMP_REMOTE_DIR/LATEST"
-DUMP_TIMESTAMP=$(cat /media/dbdump/LATEST)
+if [[ $TARGET == both ]]
+then
+	# Find latest database dump corresponding to search indexes
+
+	SEARCH_DUMP_DAY="${DUMP_TIMESTAMP/-*}"
+	"${WGET_CMD[@]}" --spider --no-remove-listing -P "$DB_DUMP_DIR" \
+		"$BASE_FTP_URL/$DB_DUMP_REMOTE_DIR"
+	DUMP_TIMESTAMP=$(
+		grep -E "\\s${SEARCH_DUMP_DAY}-\\d*" "$DB_DUMP_DIR/.listing" \
+			| sed -e 's/\s*$//' -e 's/.*\s//'
+	)
+	rm -f "$DB_DUMP_DIR/.listing"
+	echo "$DUMP_TIMESTAMP" >> "$DB_DUMP_DIR/LATEST-WITH-SEARCH-INDEXES"
+elif [[ $TARGET != search ]]
+then
+	# Just find latest database dump
+
+	"${WGET_CMD[@]}" -nd -nH -P "$DB_DUMP_DIR" \
+		"$BASE_FTP_URL/$DB_DUMP_REMOTE_DIR/LATEST"
+	DUMP_TIMESTAMP=$(cat /media/dbdump/LATEST)
+fi
 
 # Actually fetch database dump
 
-if [[ $TARGET == replica ]]
+if [[ $TARGET =~ ^(both|replica)$ ]]
 then
 	for F in MD5SUMS "${DB_DUMP_FILES[@]}"
 	do
